Move out arrow and item navigator components
