+++ site/pages/about-us/timeline-slice/navigator/index.js
@@ -1,8 +1,8 @@
 // @flow
 import React from 'react';
 import styles from './style.css';
-import Item from './item';
-import Arrow from './arrow';
+import Item from '../../../../components/navigator/item';
+import Arrow from '../../../../components/navigator/arrow';
 
 type TimelineNavProps = {
   currentIndex: number,
